# cedar-resource-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Better Metadata - Better Science",
    "version" : "2.8.3",
    "title" : "CEDAR Resource Server API"
  },
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/template-fields" : {
      "post" : {
        "tags" : [ "Template Fields" ],
        "summary" : "Create a template field",
        "description" : "Create a template field.",
        "operationId" : "createTemplateField",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "folder_id",
          "in" : "query",
          "description" : "Folder identifier. The artifact will be created in this folder. The user must have write access to the folder. If not provided, the artifact will be created in the user's home folder.",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "template_field",
          "description" : "The template field to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TemplateField"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/TemplateField"
            }
          },
          "201" : {
            "description" : "A template field",
            "schema" : {
              "$ref" : "#/definitions/TemplateField"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-fields/{template_field_id}" : {
      "get" : {
        "tags" : [ "Template Fields" ],
        "summary" : "Get a template field",
        "description" : "Get a template field.",
        "operationId" : "findTemplateField",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier. Example: https://repo.metadatacenter.org/template-fields/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A template field",
            "schema" : {
              "$ref" : "#/definitions/TemplateField"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Template Fields" ],
        "summary" : "Update a template field",
        "description" : "Update a template field.",
        "operationId" : "updateTemplateField",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folder_id",
          "in" : "query",
          "description" : "Folder identifier.",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "template_field",
          "description" : "The template field to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/TemplateField"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A template field",
            "schema" : {
              "$ref" : "#/definitions/TemplateField"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Template Fields" ],
        "summary" : "Delete a template field",
        "description" : "Delete a template field.",
        "operationId" : "deleteTemplateField",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation (no content)"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-fields/{template_field_id}/details" : {
      "get" : {
        "tags" : [ "Template Fields", "Resource Details" ],
        "summary" : "Get details of a template field",
        "description" : "Get details of a template field.",
        "operationId" : "findTemplateFieldDetails",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-fields/{template_field_id}/download" : {
      "post" : {
        "tags" : [ "Template Fields" ],
        "summary" : "Download a template field",
        "description" : "Download a template field as JSON or YAML, selected via the Accept header.",
        "operationId" : "downloadTemplateField",
        "produces" : [ "application/json", "application/x-yaml" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "Accept",
          "in" : "header",
          "description" : "Desired output format: 'application/json' or 'application/yaml'.",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "compact",
          "in" : "query",
          "description" : "When downloading YAML, produce a compact representation.",
          "required" : false,
          "type" : "boolean"
        } ],
        "responses" : {
          "200" : {
            "description" : "The template field content as an attachment"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-fields/{template_field_id}/permissions" : {
      "get" : {
        "tags" : [ "Template Fields", "Permissions" ],
        "summary" : "Get permissions of a template field",
        "description" : "Get permissions of a template field.",
        "operationId" : "getTemplateFieldPermissions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Template Fields", "Permissions" ],
        "summary" : "Update permissions of a template field",
        "description" : "Update permissions of a template field.",
        "operationId" : "updateTemplateFieldPermissions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-fields/{template_field_id}/report" : {
      "get" : {
        "tags" : [ "Template Fields", "Resource Report", "Versioning" ],
        "summary" : "Get report of a template field",
        "description" : "Get report of a template field.",
        "operationId" : "getTemplateFieldInstanceReport",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/template-fields/{template_field_id}/versions" : {
      "get" : {
        "tags" : [ "Template Fields", "Resource Report", "Versioning" ],
        "summary" : "Get a list of versions of a template field",
        "description" : "Get a list of versions of a template field.",
        "operationId" : "getTemplateFieldVersions",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_field_id",
          "in" : "path",
          "description" : "Template Field identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
    "/templates/{template_id}" : {
      "get" : {
        "tags" : [ "Templates" ],
        "summary" : "Get a template",
        "description" : "Get a template.",
        "operationId" : "findTemplate",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier. Example: https://repo.metadatacenter.org/templates/8bc64ab5-df6b-48c8-8c61-6c016245918e",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "A template",
            "schema" : {
              "$ref" : "#/definitions/Template"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Templates" ],
        "summary" : "Update a template",
        "description" : "Update a template.",
        "operationId" : "updateTemplate",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier.",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "folder_id",
          "in" : "query",
          "description" : "Folder identifier.",
          "required" : false,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "template",
          "description" : "The template to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Template"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A template",
            "schema" : {
              "$ref" : "#/definitions/Template"
            }
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Templates" ],
        "summary" : "Delete a template",
        "description" : "Delete a template.",
        "operationId" : "deleteTemplate",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Successful operation (no content)"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
    "/templates/{template_id}/details" : {
      "get" : {
        "tags" : [ "Templates", "Resource Details" ],
        "summary" : "Get details of a template",
        "description" : "Get details of a template.",
        "operationId" : "findTemplateDetails",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "template_id",
          "in" : "path",
          "description" : "Template identifier.",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not found"
          },
          "500" : {
            "description" : "Internal server error"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "api_key" : {
      "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "Template" : {
      "type" : "object",
      "properties" : {
        "@id" : {
          "type" : "string",
          "description" : "Unique URL identifier representing a specific template."
        }
      },
      "description" : "A CEDAR template (open JSON-LD document)."
    },
...
    "TemplateField" : {
      "type" : "object",
      "properties" : {
        "@id" : {
          "type" : "string",
          "description" : "Unique URL identifier representing a specific template field."
        }
      },
      "description" : "A CEDAR template field (open JSON-LD document)."
    },
...
  }
}